Modify components moved to separate namespace
