+++ src/GrasshopperTeklaDrawingLink/Components/ModifyComponents/ModifyPartComponent.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using Tekla.Structures.Drawing;
 
-namespace GTDrawingLink.Components
+namespace GTDrawingLink.Components.ModifyComponents
 {
     public class ModifyPartComponent : TeklaComponentBaseNew<ModifyPartCommand>
     {
